First commit for Salesforce library mixin for supporting page objects
This adds the keywords 'set current page object', 'go to tab' and
'current page should be'. Each is responsible for setting the "current
page object" which is just a fancy way of saying it moves that keyword
library to the front of robot's library search order.

These keywords aren't part of the base PageObject class, since they must
exist external to the page objects so that they can be used before the
first page object has been loaded.