First commit of base page object classes
The classes in ObjectPages serve two purposes. One, they should be used
as the base class for any page objects related to the the list, home and
new pages for objects. Two, when a non-existent page object is requested,
one of these classes will be instantiated and the PAGE_TYPE and
OBJECT_API_NAME attributes will be set accordingly.

Eventually we'll add similar classes for other page types, eg: the
object_recordPage page type.