Validate GitHub service attributes before saving
This involves adding a general mechanism for a service to configure a validator.
If configured, the `cci service connect` command will call the validator passing the service attributes that were input,
and abort if the validator raises an exception.

I also renamed cumulusci.core.utils.import_class to import_global since it can import anything from a module's globals.
import_class remains as an alias for backwards-compatibility.