Fix a bug in printing the flow summary when there are sub-steps.
Freezing a flow to publish to MetaDeploy can generate sub-steps where one step from the original flow is split into multiple steps.
For example update_dependencies gets split into one sub-step for each package.
The sub-steps get an extra number like ".1" added to the end of the step path, which breaks the flow summary code.
This is a quick fix to just skip trying to display the sub-steps.