Refactoring of _get_page_object and set_current_page_object
In a nutshell, previously a page object was loaded into the PageObject
library itself. Now, each page object is wrapped in a thin wrapper
to convert a class into a proper robot library. In other words, instead
of a polymorphic PageObject library, we have lots of little discreet
libraries. This makes it possible to have several libraries loaded at
once, and makes it easier for a page object to access the keywords in
some other page object. This should have been the architecture from the
start.

One big benefit is that we're no longer tweaking one of robot's private
data structures. That always felt icky, and ultimately was responsible
for a bug that triggered this refactoring. This also makes it possible to
have several page objects loaded at once.