Add cli.options module
This commit adds a options module that controls the precedence of
configuration options by means of Click callbacks. Configuration is set
from the first value found:

1. Environment variables
2. Global configuration options
3. Click option (passed from cli)
4. Click option (default value)

The two environment variables are:
- "CUMULUSCI_NO_PROMPT" :: controls prompts for expired scratch orgs.
- "CUMULUSCI_PLAIN_OUTPUT" :: controls the formatting of CliTables

Handling of the environment variables could be smarter. The type of the
option value is passed to the callback as part of the `param` argument.
This could be used to cast future non-boolean values to the correct
type. For now, any truthy value, including "false", will be treated the
same as True.