Use global callback handler
This commit revises the cli.options module to use Click.Option's built-in
handling of environment variables, and creates a single callback to
lookup global config values based on the environment variable. For
example,

- Environment Variable :: `CCI__CLI__ALWAYS_RECREATE`
- Global Config option :: `cli__always_recreate`

Also, the order of precedence is changed to:

1. Global configuration options
2. Click Option
3. Environment variables
4. Click Option default value

Using the global config as an override is the preferred approach here.
The class of options we've defined global_config values for are either
`-y`-type options or UI config flags.

This change means the justification for a separate options module is
weakened somewhat, but I believe it still makes the most sense to keep
the environment variables and callback(s) in the same place.