Add skip_types option to uninstall_packaged_incremental
This commit exposes the `self.skip_types` option as a task option. As
shown by the included test change, this is primarily intended to allow
entire metadata types to be ignored by this task. While an argument for
making explicit each ignored component can be made, in practice most
SFDO teams simply add all CustomObjectTranslation components in the
package to the `ignore` task option.

The next logical step would be to allow wildcard (`*`) matches against
component names to allow different translations of the same object to be
ignored without listing each component individually.