Rename "batches" to "subjobs" and handle corner case.
Previously, if the job were invoked in parallel (somehow) it would
be kind of random when this task would end. Now it will wait for
both processes to finish.

Job A starts and CCI is waiting on it.
Job B starts
Job A ends
CCI still waits because B was started after A and is thus "in scope".
Job B ends
CCI batch wait ends.

Job A starts
Job B starts and CCI is waiting for it.
Job A ends
CCI still waits.
Job B ends
CCI batch wait ends.

Job A starts and CCI is waiting for it.
Job B starts
Job B ends
CCI still waits because it knows it is waiting for things started since Job A.
Job A ends.
CCI batch wait ends.

Job A starts
Job B starts and CCI is waiting for it.
Job B ends
CCI batch wait ends because Job A is considered out of scope.
Job A ends.