Allow the user to specify generate_mapping_file
 * The task info implied they could, but we overwrote it with a tempfile.

Also simplify that code a bit. There was no reason to open the tempfile.

 * The code just evolved that way from previous code where it was a
TemporaryNamedFile().

Fix init_db so it is more similar to the (unfortunately not identical)
version in load.py when it comes to tables that appear in multiple
mappings.

Change generate_from_yaml to reuse Snowfakery's database handling
code instead of overriding it.

Add a try:finally to close the outputstream "just in case".

Re-indent a bunch of code.

Never save a column named RecordType to Salesforce. It's presumably
just there to represent the...record type. This prevents me
from having to use one mapping file to generate the database
tables and a different one to do the actual load.