Switch to setup-python@v2
At first glance, it appears that v1 is limited to the python versions
installed in the VM. v2 will download and install versions from
actions/python-versions. There may be some reason why we went with v1,
but it's been a while, so let's try it and see.

Also, specify the semver range for release candidates so we don't have
to do it again in a month or two.