Make sure parallel robot runs have pythonpath set properly Prior to this fix, processes spawned by pabot didn't have the current working directory in it's pythonpath. This caused relative imports to fail (eg: Resource  robot/SAL/resources/SAL.robot)
Tested with SAL, NPSP, and CaseMan. In each, parallel tests runs were
possible. Some tests fail, but they no longer fail due to this problem.
The problems seem to be related to the fact that the tests weren't
designed to be run in parallel, or they were tagged as unstable.