[wip]: Add record_types options into package xml
Tasks can call ProfileGrantAllAccess with:

```yaml
update_admin_profile:
    options:
        record_types:
                - record_type: "{managed}Foo__c.{namespaced_org}Baz"
                  default: true
```

If the referenced record type is not included in the default package.xml
the task will throw an exception unless a custom package.xml is
provided.

Instead, this commit updates the class such that any record types are
appended to the package.xml before retrieval.

FIXME: This is tagged as a `[wip]` commit because this may not be the
right way to handle namespace injection for these.