Set create_release to false by default
The create_release option doesn't play well with ref. I think we need to
build this up a bit, to decouple tags in the source repo from a created
release in the target repo. It would be better to allow a source ref to
be used to create a release in the target whether or not it has a tag.