New locators for Spring '21, plus new keyword for modal field errors 'Modal should show edit error for fields'.
This verifies that the "we hit a snag" popup lists one or more fields
which need attention. This keyword should be used instead of 'Modal should
contain errors' for pages that are using the new popup. The old keyword
is sticking around in case there are pages I don't know about which still
use the old error reporting method.