execute_anon supports minimizing Apex
The intent of this feature is to reduce the URI length for the GET Tooling API execute anonymous request by giving an option to `minimize_apex`.   This option may have unitended consequences since it alters the Apex before the GET request, and therefore, this option is off/False by default.

- `minimize_apex` option  defaults to `False` to not interfere with existing implementations.
- If `minimize_apex` option is `True`, `_prepare_apex` does the following:
    - Removes block comments
    - Removes inline comments
    - Removes whitespace at the start of lines