Add Composite Request Task
This commit adds a task wrapping the composite REST API resource. Given
a list of JSON files, this task will replace namespace and user ID
tokens, POST them and process the returned composite result. If the
request is allOrNone and an error is returned a SalesforceException is
raised, otherwise partial successes will not raise an exception.

Probable next steps:
- TODO: Add support for Composite Graphs
- TODO: Better error formatting
- TODO: Accept YAML files