Add Composite Request Task (#2341)
* Add Composite Request Task

This commit adds a task wrapping the composite REST API resource. Given
a list of JSON files, this task will replace namespace and user ID
tokens, POST them and process the returned composite result. If the
request is allOrNone and an error is returned a SalesforceException is
raised, otherwise partial successes will not raise an exception.

Probable next steps:
- TODO: Add support for Composite Graphs
- TODO: Better error formatting
- TODO: Accept YAML files

* Add task docs for CompositeApi

* Set default value for composite allOrNone property

Since allOrNone is an optional property on composite requests it may not
be present. If it isn't, then follow the API's default behavior.

* Update group for insert_record and composite_request

* Process data_files with process_list_arg

* Log composite subrequest successes/errors via logger

* Update data_files description

Make it clearer that each JSON file contains a composite request body.

* Use sf.restful instead of constructing another API client

* Use correct option lookup for _init_options

Co-authored-by: David Glick <dglick@salesforce.com>

Co-authored-by: David Reed <d.reed@salesforce.com>
Co-authored-by: David Glick <dglick@salesforce.com>