Tweaked initialization of our custom locator strategies SeleniumLibrary support isn't all that it should be - it raises an exception if you register the same locator strategy twice but doesn't provide any way to know ahead of time if a strategy has been located.
It looks like I initialized a flag (${LOCATION STRATEGIES INITIALIZED})
in the past, but never used the flag. Now the code uses it.

And, just in case you were wondering, no, we can't do this in the
__init__, since SeleniumLibrary also raises an exception if you
register the strategy before a browser is open. Finicky is thy name.

I like custom locator strategies, but their implementation could use
some improvement :-\