Handle directories missing a trailing slash
This commit tweaks how github_copy_subtree filters filenames based on an
an include list. Previously, this task would silently generate
incorrect/truncated commits when a directory was included in
self.options.includes without a trailing slash. Now, check the zipfile's
namelist's directories (filenames ending with a slash) for includes that
are missing a slash and appends it.