Add error handler decorator
Realized this should be a decorator as well. Saves the trouble of
manually formatting the string, but we have the option to do so if
callers want to handle exceptions themselves.

Also added some direct tests of the formatter, and re-arranged it to
fix a bug uncovered by said tests.