Run Feature Test workflow on PRs
This commit changes the events triggering the feature test workflow to
one of the following:

1. Manual runs:
    gh workflow run feature_test.yml
2. Pull requests
3. Commits to `main`

Previously, it compared HEAD against origin/main on every push, but that
didn't necessarily match the context; you could, for example, be in a
child branch.

Ultimately, it made the most sense for this to be triggered by
the pull_request event instead.

---

Workaround PR event "design"
----------------------------

The `actions/checkout` action will checkout `refs/pull/000/merge` which
I _think_ is the same as a merge between `refs/pullreqs/000` and the
base branch. (Apparently this is the default for travisci, but it's been
forever.)

So, instead we point checkout at the head ref (via the webhook payload)
and use fetch on the *base* ref. But that alone doesn't work, so we also
fetch both references explicitly, and use two dots instead of three.

🤬

---

Add pre-commit cache
--------------------

Add pre-commit cache and skip when cache-hit

- Move `pre-commit` install to a separate step.
- Switch to using the python version as the cache key for precommit's
  install.
- Skips install steps when there's a cache hit.
- Restrict fetch to depth=1 and skip the tags