First cut at "advanced" robot documentation
We don't have time to write a whole new section, so I took the original
docs and removed all of the information that is in the new documentation,
and then tweaked some headings and added links.

We might want to include some additional examples, such as running a
task from a keyword.  At least this way, when we public the new
documentation we don't lose any information.