Reverted a change made a few months ago.
Prior to this fix the test would  randomly fail when this suite was run
in random order. With this fix it passes every time. I'm not sure why
the code was changed to check for failure_occurred rather than
capture_page_screenshot, especially since I wanted to explictly check
for capture_page_screenshot.