Reorder github service prompts
This commit moves the call to the device authentication flow to the end,
preventing the user from having to return to the CLI and enter their
email address.

A case could be made for leaving it in place and providing the account's
default email as the default, but that approach presents a dubious value
proposition at first glance given how unpleasant passing around click
contexts sounds to me.