Compress VCR Files (#3138)
Compress VCR files using three techniques.

It looks in a small database of "known" request/response pairs to
reuse them. If it finds a hit, it creates an include_file reference.

It looks in a similar database of "known" string patterns (mostly XML) and compresses those too.

It looks for duplication within a single YAML file and uses YAML
references to de-duplicate.

Nothing for release notes.