Use GraphQL to find latest chronological release
Github sorts releases based on the tag, with tags being sorted on the
timestamp of the tagged commit, but with day granularity. Tags whose
commits are created the same day are sorted lexicographically by name
such that 9 > 10.

Here we add a new helper function which uses a chronologically sorted
GraphQL query to retrieve the tag for the latest release. Approches we
tried that didn't work:

1. Retrieving all required fields and constructing a github3 `Release`
   object using `Release.from_json`. Tried and abandoned because we'd
   have to retrieve all child nodes (i.e. `ReleaseAsset`) to do so.
2. Retrieving the release id and using `repo.release(id)` because the
   integer id needed by the REST API isn't included in the GraphQL
   `Release` object.

Retrieving `tagName` and using `repo.release_from_tag(tagName)` seemed
the easiest way to go.

I'm a bit defensive about the use of string.Template here, but it was
more readable than either concatenating strings or double bracketing
everything, but that's debatable:

```python
f"""
  query {{
    repository(owner: "{repo.owner}", name: "{repo.name}") {{
      releases(last: 1, orderBy: {{field: CREATED_AT, direction: ASC}}) {{
        nodes {{
          tagName
        }}
      }}
    }}
  }}
"""
```

- Fix tests broken by GQL