Added 'Delete records and close browser' to SalesforcePlaywright.
I'm conflicted about this - it just calls two other keywords, so a test
could easily just do something like this to avoid a custom keyword:

`run keywords  close browser  ALL  AND delete session records`

OTOH, people might forget to call that second keyword, and they are already
used to calling this one keyword in selenium-based tests. For now, and for
symmetry with the selenium keywords, having the keyword seems better than
not having the keyword.