Add custom hosted runners
This commit replaces GitHub-hosted runners with custom-hosted ("premium"
in internal documentation) runners with known IP ranges.

The new runners mapped to old labels:

| New                   | Old              | Operating System    |
|-----------------------|------------------|---------------------|
| `sfdc-ubuntu-latest`  | `ubuntu-latest`  | Ubuntu 20.04 LTS    |
| `sfdc-windows-latest` | `windows-latest` | Windows Server 2022 |

The new `runs-on` entries map to labels, and should be stable.

TODO: Determine whether the image used for custom-hosted contains the
same software as the default ubuntu-latest