Include default SFDX gitignore files
We previously didn't include the `.sf` directory in our gitignore
template, which resulted in git tracking it. Because `.sf` includes the
git_dir used in source tracking this resulted in a large number of
changes that users did not expect.

Because our copy of the gitignore is pretty old, it didn't include any
of the entries included in the skeleton generated by `sfdx
force:project:create`.

We now inline the contents of entire template [source] in our own.

TODO: We probably need to revisit our DX integration test plan.

[source]: https://github.com/forcedotcom/salesforcedx-templates/blob/main/packages/templates/src/templates/project/gitignore