Skip some tests on firefox until we can solve the root problem.
The root problem is that firefox won't render some of the object manager
pages, and instead shows "We can't display this page because your browser
blocks cross-domain cookies, but you can view this page in Salesforce
Classic. Click here to open this page in Salesforce Classic."

There's a firefox setting to turn off the cross-domain cookie checks,
but at the moment I don't know how to set that when opening the browser.