I revamped how we pass some CumulusCI context to javascript keywords
This adds the path to the CumulusCI node_modules folder to NODE_PATH,
along with a `cumulusci.js` module that is initialized by the robot
task and jsextension loader. With this, keyword files can import
the cumulusci module to get access or the org name, instance url,
and ord id, along with the repo name and repo root.