Add new keyword "Wait until salesforce is ready" This uses `Wait until loading is complete` but also adds an additional check for landing on a classic page.
This also adds a new 'wait' parameter to the playwright version of
`Open test browser`, for parity with the selenium version. Plus, it makes
testing this particular feature easier.

The wait times are admittedly a bit arbitrary, but they seem to work. If
I reduce them much, they will fail more often when Salesforce is running
slower than usual.

I don't recall if I'm still supposed to check in the documentation or if
we regenerate it when we regenerate the rest of the docs, so I'm including
it here.