Ok, I'm giving up on `wait_until_network_is_idle`.
On the surface this keyword seems like the logically correct thing to
do, but it doesn't behave the way I expect on Salesforce. When it times
out, a screenshot shows a fully rendered page so I don't understand why
it thinks the network is busy.

Using a `sleep` here is exactly what we do in the selenium world, so if
it's good enough there it's good enough here.