Fix test that broke in spring '23.
The test was using a salesforce page that did not use lightning components,
but that page has been rewritten to use lightning componets. With this
fix, instead of testing the keyword with a salesforce page, it uses
a static html file. The goal of the test is to make sure form keywords
can input data in non-lightning pages, and a static page works fine for
that purpose. We already had this static file being used for some other
tests, but I needed to tweak it slightly for this test.

Also, I renamed a couple of tests to remove the object type from the name,
since that information is somewhat irrelevant.