Refresh all objects in org schema cache (#3473)
Previously the org schema was versioned as a unified whole. Not one sobject at a time. This worked until I added the feature of allowing one to download a subset of the schema. Now some objects might be refreshed on Monday, some on Tuesday, a different set on Wednesday etc. To allow this, all objects are independently versioned (with a last_modified_date) now. If it is needed by the application it is refreshed and if it isn't, it isn't.

This bug was hidden because of the ways we were using get_org_schema before we added capture_sample_data. Now we really need the schema to be complete and up-to-date.

Tasks that use get_org_schema include load_data (for the Recently Used checking) and generate_mapping.