Remove deprecated Playwright keywords
We missed a major release of `robotframework-browser` ([v15]), which
removed the deprecated `execute_javascript` keyword. This resulted in
test failures that we initially believed were timeout related because
the `AttributeError` was being cought by the bare `except` and re-raised
with an misleading error message.

This commit resolves the falures by:

1. Switching to the `evaluate_javascript` keyword
2. Raising the timeout exception using the `raise ... from exc`

[v15]: https://github.com/MarketSquare/robotframework-browser/releases/tag/v15.0.0