Fix sdist and test that it can be installed (#3501, #3520)
Per the title, verify that any built artifacts can be installed.

When I tried to install the wheel from main, they would not install due to a typo which I separately reported. "CumulusCI-Test"

This reminded me that because I have run into similar problems in Snowfakery, I have a build there that tests that it can be installed.

Unless I am mistaken, the line that is failing in this test is

                  pip show cumulusci
In other word, it i proving that the main branch does not build a wheel that can be installed and inspected under the name cumulusci

If my hypothesis is correct, fixing that bug on main will also fix this PR and then it can be merged.

Or...it may reveal other problems. I think there may be a problem with the sdist.

---

* Include readme fragments in sdist ()

If not included the sdist will not be installable, and will fail with:

```
hatch_fancy_pypi_readme.exceptions.ConfigurationError:
["Fragment file 'README.md' not found.", "Fragment file 'docs/history.md' not found."]
```

---------

Co-authored-by: James Estevez <jestevez@salesforce.com>