Fix skip_future_releases option
Fixes #3594

As written, skip_future_releases is always `True` for
`skip_future_releases: False` or `skip_future_releases: false` because
those get converted to booleans before being passed to
`process_bool_arg`. This was working as designed when run through the
CLI because the option doesn't go through the Pydantic model so it's a
`str`.