Task to retrieve a complete Profile from an org (#3672)
[W-8932343](https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000097PhOIAU/view)
- Build a task to retrieve a complete Profile from an org

Permissionable Entities Retrieved:
1. CustomApplication
2. ApexClass
3. CustomPermission
4. CustomObject
5. ExternalDataSource
6. Flow
7. ApexPage
8. CustomTab

Issues:
1. I'm able to retrieve flows within the same package.xml, but they are
not visible under the Profile metadata. **(Added functionality to
explicitly add them to the profile-meta.xml file)**
2. While retrieving apps, those without a StartUrl in the AppMenuItem
table are not being included in the retrieval process. This seems to be
a package.xml issue, not something within our control.
3. Two sObjects, AIRecordInsight and Document, are not being
successfully retrieved even when specified in the package.xml. **(Edit:
However, they show up under the ObjectPermissions in the
profile-meta.xml file, so no issues)**

---------

Co-authored-by: David Reed <d.reed@salesforce.com>