Fix `is_person_type` `AttributeError` in record type handling (#3739)
Fix for
[issue](https://trailhead.salesforce.com/trailblazer-community/feed/0D54V00007XIPZMSA5)
related to your recent PR
([3702](https://github.com/SFDO-Tooling/CumulusCI/pull/3702)).

**Error Message:** `AttributeError: is_person_type`

**Issue Reason:**
During load, the tasks expects the column `is+person_type` to be present
in the record type table `rt_mapping_table`. This column is present for
sql files which do extract after the new release of CumulusCI (v3.84.0)
but will not be present for existing sql files.

**Fix:**
Previously the `outerjoins_to_add` function of the
`AddRecordTypesToQuery` class checks if the combination of
`is_person_type` and `developer_name` is unique. Now, if the
`rt_mapping_table` does not have the `is_person_type` table, the task
ignores the condition to check for record types based on
`is_person_type` and only checks for `developer_name` (as it was
previous to this PR
[3702](https://github.com/SFDO-Tooling/CumulusCI/pull/3702))

---------

Co-authored-by: Jaipal Reddy Kasturi <jkasturi@salesforce.com>