ADR: Migrate from SFDX to SF CLI (#3630)
Salesforce has announced the deprecation of SFDX (v7) and is now
focusing on the newer SF CLI (v2). They've said that SF CLI (v2) is
smart enough to understand all SFDX commands, as well as new SF
commands. However, our web applications and other integrations depend on
SFDX. SFDX is end-of-life, so we need to cut over and make some
improvements by using the new, better features. This ADR proposes that
we migrate to the SF CLI.