fix: Polymorphic extraction filtering
Running extraction against a target org resulted in a ConfigError from a
mapping mismatch. It's not clear to me why this works, but by filtering
the offending mapping.sf_object against the OPT_IN_ONLY list does the
trick. Best guess is that the latest changes sidestepped the earlier
filtering logic in some respect, but I don't understand bulk_data enough
to say for certain.

This fix cleared the immediate error, but another related to missing
schema has shown up, so this may not be the correct fix