@W-15946941: New task added to display files in the salesforce org  (#3799)
Files uploaded via the app launcher in the orgs are currently not
included in the metadata listings.

To address this, the "display_files" task has been introduced to list
these files efficiently.

- This task gives the list of files in a tabular format querying the
`ContentDocument`, fetching the fields Id, Filetype, and Filename, as
filenames alone are not unique identifiers.
- Upon uploading a new file, a fresh entry is generated in the
`ContentDocument` entity, establishing it as the latest version in the
`ContentVersion` entity. Subsequent uploads of new versions of already
tracked documents are noted in the ContentVersion entity, without
creating additional ContentDocument records.
- Retrieving the latest version involves filtering the ContentVersion
entity by setting IsLatest to true and employing the file's
ContentDocumentId as the filter criterion.