Update the validation for loading to check the required fields as well (#3807)
To ensure field-level permissions and successful dataset load, there are
several validations performed on the `mapping.yml` file. These
validations check whether the fields and SObjects have the required
permissions, and whether namespaces need to be injected, as well as
handling case insensitivity for the `mapping.yml`. This functionality is
already implemented in the function `validate_and_inject_mapping`.

However, there was a missing corner case where the function did not
capture errors for required fields missing in the mapping.yml. This
functionality has now been added, and the function is used for the
preflight check.

---------

Co-authored-by: James Estevez <jestevez@salesforce.com>