migrate entrypoints handling to standard library importlib.metadata (#3904)
Declaring a namespace package has gone through a few revisions.
pkg_resources has a version that is heavily deprecated. pkgutil provides
a python2/python3 compatible version that is also compatible with native
python3 namespaces.


https://packaging.python.org/en/latest/guides/packaging-namespace-packages/

pkg_resources is very very deprecated and importing or using it results
in deprecation warnings. It's time to move off of it entirely.

---------

Signed-off-by: Rupesh J <rupesh.j@salesforce.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Arvind Bhamidipati <arvind.bhamidipati@salesforce.com>
Co-authored-by: James Estevez <jestevez@salesforce.com>
Co-authored-by: Bharath Chadarajupalli <vchadarajupalli@salesforce.com>